requirements.txt
python-telegram-bot[job-queue]>=20.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
Pillow>=10.0.0